Refactor: Use poetry (#106)
* chore(git): Ignore JetBrains IDEA configuration directory

* refactor(python): Use Poetry for package management

Advantages:

- Separates development and production dependencies.
- The dependency version is specified separately from the lock file. In
practice this means that the version in pyproject.toml generally only
needs to be set to anything other than "*" if and when it becomes
necessary to use a specific version range.
- The lock file includes SHA-256 checksums by default, and these are
checked during installation.

Disadvantages:

- More complex configuration than Pip.
- Python package managers come and go, and this one is likely going to
suffer the same fate *eventually.*
- Introduces poetry.toml simply to specify that the virtualenv should be
in the project directory. The default is to put virtualenvs in
~/.poetry, which is a non-standard location and therefore might
interfere with typical IDE setups, mounting the virtualenv in containers
or VMs, and the like.

CI installation based on
<https://github.com/python-poetry/poetry/blob/master/.github/workflows/main.yml>.