feat: Support only recent Python versions
The template repo originally supported Python 3.6 onwards, which is a
bit limiting since packages like Black have removed support for 3.6
already. I'm guessing for the majority of projects we'd prefer them to
start with the latest stable version of Python, to have the most useful
set of features and packages. Only if necessary would users change the
supported range of Python versions.