feat: Create single summary test job (#237)
GitHub branch protection rules do not have a feature for specifying that
*all* workflow jobs must pass before the branch can be merged. Instead
we have to specify the job names explicitly. Since each job in a matrix
gets a different name, that means whenever the matrix is updated the
branch protection rule must be updated. To work around this we use a
non-matrix "finalise" job which checks that all dependent jobs passed.

Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>