build: Use literally latest runner instead of locking to latest
This is a compromise between reliability (`ubuntu-latest` could break
things whenever it changes) on the one hand and maintainability (not
having to change the code very often) and detecting issues with the
latest runner ASAP on the other hand. Since this repo isn't deployed
anywhere, there's no risk of an "all hands on deck" situation to salvage
the situation if `ubuntu-latest` breaks things, it's just going to be
fixed when we have the time. Included a comment to dissuade any users
from using `ubuntu-latest` in instantiations of this template.